refactor(peer): split the overlay.go god file by responsibility (#900)
Phase D (part 3) — overlay.go was a 2.7k-line god file mixing lifecycle,
inbound handshake, the event loop, message handlers, broadcast/relay
fan-out and RPC JSON shaping. Split the natural seams into siblings in the
same package (pure code move, no behaviour change):

- overlay_inbound.go: the accept/handshake path (startListener, acceptLoop,
  handleInbound, performInboundHandshake, handshakeConfigFor,
  chargeInboundHandshake) + the inbound tuning consts.
- overlay_broadcast.go: the broadcast/relay fan-out (Broadcast family,
  forEachConnected, sendAndLog, RelayFromValidator) and the relayed-index
  reverse map.
- overlay_rpc.go: the `peers` RPC JSON shaping (PeersJSON, ClusterJSON,
  nodeStatusRPCName).

overlay.go drops from 2722 to 2007 lines.

Refs #900
