+++ internal/peermanagement/overlay.go
@@ -1,7 +1,6 @@
 package peermanagement
 
 import (
-	"bufio"
 	"bytes"
 	"context"
 	"crypto/rand"
@@ -11,14 +10,11 @@ import (
 	"fmt"
 	"log/slog"
 	"net"
-	"net/http"
-	"strconv"
 	"strings"
 	"sync"
 	"sync/atomic"
 	"time"
 
-	addresscodec "github.com/LeJamon/go-xrpl/codec/addresscodec"
 	"github.com/LeJamon/go-xrpl/internal/peermanagement/cluster"
 	"github.com/LeJamon/go-xrpl/internal/peermanagement/message"
 	"github.com/LeJamon/go-xrpl/internal/peermanagement/peertls"
@@ -26,16 +22,6 @@ import (
 	"github.com/LeJamon/go-xrpl/protocol"
 	"golang.org/x/sync/errgroup"
 )
-
-// inboundBacklogSlack caps the accept-side goroutine count to
-// MaxInbound + slack so a burst of accepts cannot fan out unbounded;
-// canAcceptInbound is the authoritative slot gate.
-const inboundBacklogSlack = 8
-
-// acceptBackoff throttles the retry rate when listener.Accept returns
-// a non-fatal error (typically EMFILE-class) so the loop does not
-// spin at CPU speed under FD pressure.
-const acceptBackoff = 100 * time.Millisecond
 
 // serveWorkerCount bounds concurrent heavy serve operations (fetch-pack /
 // get-objects / tx back-fill) handled off the event loop, and
@@ -46,28 +32,6 @@ const (
 	serveWorkerCount = 4
 	serveQueueDepth  = 64
 )
-
-// RelayedIndexTTL bounds how long a suppression-key → peers entry is
-// kept in the reverse index. Must match the consensus router's
-// messageDedupTTL so that a hash remains queryable for as long as the
-// router may observe duplicates for it. If the index expired before
-// the dedup window, a duplicate hitting router.handleProposal could
-// find no "peers that have the message" entry and under-feed the
-// slot — the exact bug B3 was filed to fix.
-const RelayedIndexTTL = 30 * time.Second
-
-// RelayedIndexMaxEntries caps memory for the reverse index under
-// adversarial traffic. Sized to match the adaptor's dedup cap so both
-// age out together under sustained churn.
-const RelayedIndexMaxEntries = 4096
-
-// relayedEntry is one bucket in the reverse index — the set of peers
-// we know "have" a given suppression-key, plus the last-update time
-// for TTL reaping.
-type relayedEntry struct {
-	peers  map[PeerID]struct{}
-	seenAt time.Time
-}
 
 // Overlay is the central orchestrator for XRPL peer-to-peer networking.
 // It manages peer connections, discovery, message routing, and the reduce-relay system.
@@ -481,24 +445,6 @@ func (o *Overlay) IncPeerBadData(peerID PeerID, reason string) uint32 {
 	return peer.IncBadData(reason)
 }
 
-// chargeInboundHandshake charges the inbound endpoint's resource Consumer
-// for a malformed or abusive handshake. During the handshake the peer is
-// not yet in o.peers (addPeer runs only after a successful handshake), so
-// routing the charge through IncPeerBadData / the peer map would silently
-// no-op. We charge the endpoint Consumer directly by address, mirroring
-// rippled which charges the inbound endpoint's Resource::Consumer for
-// handshake abuse. The Consumer's balance persists in the manager keyed by
-// address, so a host spamming malformed handshakes accrues balance across
-// attempts and is eventually throttled at admission.
-func (o *Overlay) chargeInboundHandshake(addr, reason string) {
-	if o.resourceManager == nil {
-		return
-	}
-	c := o.resourceManager.NewInboundEndpoint(addr)
-	c.Charge(chargeForReason(reason), reason)
-	c.Release()
-}
-
 // peerNegotiatedLedgerReplay reports whether the peer identified by
 // peerID advertised the ledger-replay feature during handshake. Used
 // to gate serving mtREPLAY_DELTA_REQ and mtPROOF_PATH_REQ: these
@@ -828,274 +774,6 @@ func (o *Overlay) Stop() error {
 	})
 
 	return nil
-}
-
-// startListener creates and starts the TCP/TLS listener.
-func (o *Overlay) startListener() error {
-	tcpListener, err := net.Listen("tcp", o.cfg.ListenAddr)
-	if err != nil {
-		return err
-	}
-
-	certPEM, keyPEM, err := o.identity.TLSCertificatePEM()
-	if err != nil {
-		tcpListener.Close()
-		return fmt.Errorf("overlay: build TLS cert: %w", err)
-	}
-
-	l := peertls.NewListener(tcpListener, &peertls.Config{
-		CertPEM: certPEM,
-		KeyPEM:  keyPEM,
-	})
-	o.listenerMu.Lock()
-	o.listener = l
-	o.listenerMu.Unlock()
-	return nil
-}
-
-// acceptLoop accepts incoming connections. acceptBackoff throttles
-// retries under EMFILE-class errors; inboundSem caps the handler
-// goroutine fan-out.
-func (o *Overlay) acceptLoop(ctx context.Context) error {
-	for {
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		default:
-		}
-
-		conn, err := o.listener.Accept()
-		if err != nil {
-			if ctx.Err() != nil {
-				return ctx.Err()
-			}
-			// A closed listener is terminal — exit instead of
-			// spinning the backoff. Also the !cgo peertls stub path,
-			// which closes the inner listener at NewListener.
-			if errors.Is(err, net.ErrClosed) {
-				return err
-			}
-			select {
-			case <-ctx.Done():
-				return ctx.Err()
-			case <-time.After(acceptBackoff):
-			}
-			continue
-		}
-
-		select {
-		case o.inboundSem <- struct{}{}:
-		case <-ctx.Done():
-			conn.Close()
-			return ctx.Err()
-		}
-
-		o.peerWG.Add(1)
-		go func(c net.Conn) {
-			defer o.peerWG.Done()
-			defer func() { <-o.inboundSem }()
-			o.handleInbound(ctx, c)
-		}(conn)
-	}
-}
-
-func (o *Overlay) handleInbound(ctx context.Context, conn net.Conn) {
-	defer func() {
-		if r := recover(); r != nil {
-			slog.Error("Panic in inbound handler", "t", "Overlay", "panic", r)
-			conn.Close()
-		}
-	}()
-
-	// The inbound slot limit is enforced after the handshake (see
-	// hasInboundSlot below) because reserved/cluster peers are admitted beyond
-	// the cap and their node key is unknown until the handshake completes.
-	// Concurrent handshakes stay bounded by inboundSem regardless.
-	remoteAddr := conn.RemoteAddr().String()
-	endpoint, _ := ParseEndpoint(remoteAddr)
-
-	peerID := PeerID(o.nextID.Add(1))
-	peer := NewPeer(peerID, endpoint, true, o.identity, o.events)
-	peer.SetDroppedEventsCounter(&o.droppedEvents)
-	if err := peer.AcceptConnection(conn); err != nil {
-		slog.Warn("Inbound rejected: peer not in disconnected state",
-			"t", "Overlay", "remote", remoteAddr, "err", err)
-		conn.Close()
-		return
-	}
-
-	tlsConn, ok := conn.(peertls.PeerConn)
-	if !ok {
-		slog.Error("Inbound connection is not peertls", "t", "Overlay", "remote", remoteAddr)
-		conn.Close()
-		return
-	}
-
-	if err := o.performInboundHandshake(ctx, peer, tlsConn); err != nil {
-		slog.Info("Inbound handshake failed", "t", "Overlay", "remote", remoteAddr, "err", err)
-		conn.Close()
-		o.dispatchLifecycle(Event{
-			Type:     EventPeerFailed,
-			PeerID:   peerID,
-			Endpoint: endpoint,
-			Inbound:  true,
-			Error:    err,
-		})
-		return
-	}
-
-	if o.isConnectedTo(endpoint) {
-		conn.Close()
-		return
-	}
-
-	if !o.hasInboundSlot(peer) {
-		slog.Info("Inbound rejected: no slots", "t", "Overlay", "remote", remoteAddr)
-		conn.Close()
-		return
-	}
-
-	peer.setState(PeerStateConnected)
-	slog.Info("Inbound peer connected", "t", "Overlay", "remote", remoteAddr)
-
-	o.addPeer(peer)
-
-	o.peerWG.Add(1)
-	go func() {
-		defer o.peerWG.Done()
-		err := peer.Run(ctx)
-		if err != nil {
-			slog.Info("Inbound peer run ended", "t", "Overlay", "remote", remoteAddr, "err", err)
-			o.notePeerRunEnded(err)
-		}
-		o.removePeer(peerID)
-	}()
-}
-
-func (o *Overlay) performInboundHandshake(ctx context.Context, peer *Peer, tlsConn peertls.PeerConn) error {
-	// The peer is not in o.peers during the handshake, so bad-data
-	// charges route through the endpoint Consumer keyed by this address.
-	addr := peer.Endpoint().String()
-
-	// Accept() does not drive the handshake; complete it before reading
-	// the Finished bytes for SharedValue.
-	handshakeCtx, cancel := context.WithTimeout(ctx, o.cfg.HandshakeTimeout)
-	defer cancel()
-	if err := tlsConn.HandshakeContext(handshakeCtx); err != nil {
-		return NewHandshakeError(peer.Endpoint(), "tls", err)
-	}
-
-	sharedValue, err := tlsConn.SharedValue()
-	if err != nil {
-		return NewHandshakeError(peer.Endpoint(), "shared_value", err)
-	}
-
-	deadline := time.Now().Add(o.cfg.HandshakeTimeout)
-	tlsConn.SetDeadline(deadline)
-	defer tlsConn.SetDeadline(time.Time{})
-
-	bufReader := bufio.NewReader(tlsConn)
-	req, err := http.ReadRequest(bufReader)
-	if err != nil {
-		return NewHandshakeError(peer.Endpoint(), "read_request", err)
-	}
-	req.Body.Close()
-
-	// Server-Domain runs first in the verify chain.
-	if _, err := ValidateServerDomain(req.Header); err != nil {
-		o.chargeInboundHandshake(addr, "handshake-malformed-extras")
-		return NewHandshakeError(peer.Endpoint(), "verify_extras", err)
-	}
-
-	// Build the handshake config once and share it with both
-	// VerifyPeerHandshake and BuildHandshakeResponse so the inbound
-	// and outbound paths cannot diverge.
-	hsCfg := o.handshakeConfigFor()
-
-	// Full session-signature verification — the whole point of #269.
-	peerPubKey, verifyErr := VerifyPeerHandshake(
-		req.Header,
-		sharedValue,
-		o.identity.EncodedPublicKey(),
-		hsCfg,
-	)
-	if verifyErr != nil {
-		if !errors.Is(verifyErr, ErrSelfConnection) && !errors.Is(verifyErr, ErrNetworkMismatch) {
-			o.chargeInboundHandshake(addr, "handshake-verify")
-		}
-		return NewHandshakeError(peer.Endpoint(), "verify", verifyErr)
-	}
-	peer.mu.Lock()
-	peer.remotePubKey = peerPubKey
-	peer.mu.Unlock()
-
-	peerRemote := tcpRemoteIP(tlsConn)
-	extras, extraErr := ParseHandshakeExtras(
-		req.Header,
-		o.cfg.PublicIP,
-		peerRemote,
-	)
-	if extraErr != nil {
-		o.chargeInboundHandshake(addr, "handshake-malformed-extras")
-		return NewHandshakeError(peer.Endpoint(), "verify_extras", extraErr)
-	}
-	peer.applyHandshakeExtras(extras)
-
-	caps := NewPeerCapabilities()
-	caps.Features = ParseProtocolCtlFeatures(req.Header)
-	protocol := NegotiateProtocolVersion(req.Header.Get(HeaderUpgrade))
-	if protocol == "" {
-		o.chargeInboundHandshake(addr, "handshake-protocol-negotiation")
-		// Write a 400 Bad Request back so a misconfigured peer sees
-		// the rejection reason instead of a TCP RST. Best-effort: a
-		// write error here is shadowed by the negotiation failure we
-		// are already returning.
-		var remoteAddr string
-		if peerRemote != nil {
-			remoteAddr = peerRemote.String()
-		}
-		errResp := BuildHandshakeErrorResponse(
-			hsCfg.UserAgent,
-			remoteAddr,
-			"Unable to agree on a protocol version",
-		)
-		_ = errResp.Write(tlsConn)
-		return NewHandshakeError(peer.Endpoint(), "verify",
-			fmt.Errorf("%w: unable to agree on a protocol version (peer offered %q)",
-				ErrInvalidHandshake, req.Header.Get(HeaderUpgrade)))
-	}
-
-	peer.mu.Lock()
-	peer.bufReader = bufReader
-	peer.capabilities = caps
-	peer.protocolVersion = protocol
-	peer.mu.Unlock()
-
-	resp := BuildHandshakeResponse(o.identity, sharedValue, hsCfg, protocol)
-	addAddressHeaders(resp.Header, hsCfg, peerRemote)
-	if err := resp.Write(tlsConn); err != nil {
-		return NewHandshakeError(peer.Endpoint(), "send_response", err)
-	}
-
-	return nil
-}
-
-// handshakeConfigFor builds the per-handshake config used by both
-// inbound and outbound paths so they cannot drift.
-func (o *Overlay) handshakeConfigFor() HandshakeConfig {
-	return HandshakeConfig{
-		UserAgent:           o.cfg.UserAgent,
-		NetworkID:           o.cfg.NetworkID,
-		CrawlPublic:         false,
-		EnableLedgerReplay:  o.cfg.EnableLedgerReplay,
-		EnableCompression:   o.cfg.EnableCompression,
-		EnableVPReduceRelay: o.cfg.EnableVPReduceRelay,
-		EnableTxReduceRelay: o.cfg.EnableTxReduceRelay,
-		InstanceCookie:      o.instanceCookie,
-		ServerDomain:        o.cfg.ServerDomain,
-		PublicIP:            o.cfg.PublicIP,
-		LedgerHintProvider:  o.ledgerHintProviderSnapshot(),
-	}
 }
 
 // eventLoop processes internal events. It drains both the dedicated
@@ -1948,253 +1626,6 @@ func (o *Overlay) Connect(addr string) error {
 	return nil
 }
 
-// sendAndLog sends msg to peer and logs a failure under opName:
-// ErrSendBufferFull at Warn (silent drops masked TMTransaction relay loss
-// in #401), other failures at Info. Shared by the broadcast / relay
-// fan-out loops below.
-func (o *Overlay) sendAndLog(peer *Peer, msg []byte, opName string) {
-	if err := peer.Send(msg); err != nil {
-		level := slog.LevelInfo
-		if errors.Is(err, ErrSendBufferFull) {
-			level = slog.LevelWarn
-		}
-		slog.Log(context.Background(), level, opName+" send failed",
-			"t", "Overlay",
-			"peer", peer.ID(),
-			"frame_size", len(msg),
-			"err", err.Error(),
-		)
-	}
-}
-
-// forEachConnected sends msg to every connected peer for which skip
-// returns false (skip nil = no filter), logging Send failures under
-// opName. Holds peersMu.RLock for the iteration and returns the peer IDs
-// the frame was handed to (best-effort: included even when the Send
-// errored, matching the reverse-index contract). Extracts the
-// send-and-log fan-out shared by Broadcast / BroadcastExcept /
-// BroadcastExceptSet / RelayFromValidator.
-func (o *Overlay) forEachConnected(msg []byte, opName string, skip func(PeerID, *Peer) bool) []PeerID {
-	o.peersMu.RLock()
-	defer o.peersMu.RUnlock()
-
-	var sent []PeerID
-	for id, peer := range o.peers {
-		if peer.State() != PeerStateConnected {
-			continue
-		}
-		if skip != nil && skip(id, peer) {
-			continue
-		}
-		o.sendAndLog(peer, msg, opName)
-		sent = append(sent, id)
-	}
-	return sent
-}
-
-// Broadcast sends a message to all connected peers, unfiltered. Used
-// for SELF-originated validator traffic (our own proposals and
-// validations) and for non-validator messages (statusChange, etc.).
-// The squelch filter is deliberately skipped for self-originated
-// broadcasts; otherwise a peer that squelches our own pubkey would
-// silence us to them.
-//
-// For peer-originated validator messages that need to be gossip-
-// forwarded, use RelayFromValidator which applies the squelch filter
-// and excludes the originating peer.
-func (o *Overlay) Broadcast(msg []byte) error {
-	o.forEachConnected(msg, "broadcast", nil)
-	return nil
-}
-
-// BroadcastExcept sends a message to every connected peer except the
-// one identified by exceptPeer. Used for gossip of peer-originated
-// messages that are NOT per-validator (manifests) — the per-validator
-// squelch filter in RelayFromValidator doesn't apply. Pass 0 for
-// exceptPeer to fall through to a plain Broadcast.
-func (o *Overlay) BroadcastExcept(exceptPeer PeerID, msg []byte) error {
-	o.forEachConnected(msg, "broadcast-except", func(id PeerID, _ *Peer) bool {
-		return id == exceptPeer
-	})
-	return nil
-}
-
-// BroadcastExceptSet sends a message to every connected peer whose
-// ID is not present in excluded. Used by tx-set acquire to skip peers
-// that have repeatedly returned non-progressing TMLedgerData responses.
-// This is a go-xrpl-specific outbound filter; rippled does NOT remove
-// such peers from its peer set — it charges them and lets the global
-// resource manager throttle them, so the peer stays eligible for the
-// next broadcast. go-xrpl has no equivalent per-message resource accounting
-// today, hence the explicit per-acquire exclusion. A nil or empty
-// excluded map falls through to a plain Broadcast. Issue #420.
-//
-// Issue #724: the exclusion must never starve the broadcast. If every
-// connected peer is excluded, the message would reach no one and the
-// caller (tx-set missing-node acquisition) wedges in wrongLedger until
-// the TTL sweep — the recurring under-load validation stall. When that
-// happens, fall back to broadcasting to all connected peers, restoring
-// rippled's "peer stays eligible for the next request" semantics rather
-// than dropping the request on the floor.
-func (o *Overlay) BroadcastExceptSet(excluded map[PeerID]bool, msg []byte) error {
-	if len(excluded) == 0 {
-		return o.Broadcast(msg)
-	}
-	// #724: if excluding would reach no one, ignore the exclusion entirely
-	// rather than dropping the request on the floor.
-	ignoreExclusion := o.allConnectedExcluded(excluded)
-	o.forEachConnected(msg, "broadcast-except-set", func(id PeerID, _ *Peer) bool {
-		return !ignoreExclusion && excluded[id]
-	})
-	return nil
-}
-
-// allConnectedExcluded reports whether every connected peer is in the
-// excluded set (and there is at least one connected peer) — the
-// starvation condition BroadcastExceptSet overrides.
-func (o *Overlay) allConnectedExcluded(excluded map[PeerID]bool) bool {
-	o.peersMu.RLock()
-	defer o.peersMu.RUnlock()
-
-	connected, eligible := 0, 0
-	for id, peer := range o.peers {
-		if peer.State() != PeerStateConnected {
-			continue
-		}
-		connected++
-		if !excluded[id] {
-			eligible++
-		}
-	}
-	return eligible == 0 && connected > 0
-}
-
-// RelayFromValidator forwards a peer-originated validator message
-// (proposal or validation) to other connected peers, applying the
-// per-peer squelch filter on the ORIGINATING validator's pubkey AND
-// excluding the originating peer (exceptPeer). Pass 0 for exceptPeer
-// when no peer should be excluded (e.g. tests that synthesize a relay).
-//
-// suppressionHash is the consensus-router suppression key for this
-// message (same [32]byte used by the dedup cache). Every peer we
-// actually send to is recorded in the reverse index so a later
-// duplicate arrival from ANOTHER peer can query
-// Overlay.PeersThatHave(suppressionHash) and feed the reduce-relay
-// slot with the full set of known-havers.
-//
-// The squelch is consulted before each outbound send and expired
-// squelches auto-clear via Peer.ExpireSquelch. Self-origin is handled
-// by a separate code path (see Broadcast) that skips the filter
-// entirely.
-func (o *Overlay) RelayFromValidator(validator []byte, suppressionHash [32]byte, exceptPeer PeerID, msg []byte) error {
-	// forEachConnected returns the peers we forwarded to (best-effort,
-	// including any whose Send errored). Record into the reverse index
-	// AFTER it releases peersMu so we never nest index-mutex inside
-	// peers-mutex.
-	forwarded := o.forEachConnected(msg, "relay-from-validator", func(id PeerID, peer *Peer) bool {
-		return id == exceptPeer || !peer.ExpireSquelch(validator)
-	})
-
-	if len(forwarded) > 0 {
-		o.recordRelayedPeers(suppressionHash, forwarded)
-	}
-	return nil
-}
-
-// recordRelayedPeers adds peerIDs to the reverse-index bucket for
-// suppressionHash, trimming expired buckets if we hit the size cap.
-// Safe for concurrent callers.
-func (o *Overlay) recordRelayedPeers(suppressionHash [32]byte, peerIDs []PeerID) {
-	if o.relayedIndex == nil {
-		return
-	}
-	clock := o.clockForIndex
-	if clock == nil {
-		clock = time.Now
-	}
-	now := clock()
-
-	o.relayedIndexMu.Lock()
-	defer o.relayedIndexMu.Unlock()
-
-	// Trim if we're at capacity. A cheap TTL sweep rather than a
-	// formal LRU — the index is a cache, not a hot path.
-	if len(o.relayedIndex) >= RelayedIndexMaxEntries {
-		cutoff := now.Add(-RelayedIndexTTL)
-		for h, e := range o.relayedIndex {
-			if e.seenAt.Before(cutoff) {
-				delete(o.relayedIndex, h)
-			}
-		}
-		// If that didn't free enough space (adversarial churn), drop
-		// half the map — bounded worst case, same shape as the
-		// messageSuppression eviction in the consensus router.
-		if len(o.relayedIndex) >= RelayedIndexMaxEntries {
-			i := 0
-			for h := range o.relayedIndex {
-				if i >= RelayedIndexMaxEntries/2 {
-					break
-				}
-				delete(o.relayedIndex, h)
-				i++
-			}
-		}
-	}
-
-	entry, ok := o.relayedIndex[suppressionHash]
-	if !ok {
-		entry = &relayedEntry{peers: make(map[PeerID]struct{})}
-		o.relayedIndex[suppressionHash] = entry
-	}
-	for _, id := range peerIDs {
-		entry.peers[id] = struct{}{}
-	}
-	entry.seenAt = now
-}
-
-// PeersThatHave returns the set of peer IDs known to have the message
-// whose suppression-hash is `suppressionHash`. Entries are populated
-// when we relay a validator message outward (RelayFromValidator) and
-// expire after RelayedIndexTTL.
-//
-// Returns nil when the hash is unknown or the bucket has aged out —
-// callers treat both equivalently (nothing to feed the slot with
-// beyond the current originPeer).
-//
-// Thread-safe. The returned slice is a private copy the caller may
-// mutate freely.
-func (o *Overlay) PeersThatHave(suppressionHash [32]byte) []PeerID {
-	if o.relayedIndex == nil {
-		return nil
-	}
-	clock := o.clockForIndex
-	if clock == nil {
-		clock = time.Now
-	}
-
-	o.relayedIndexMu.Lock()
-	defer o.relayedIndexMu.Unlock()
-
-	entry, ok := o.relayedIndex[suppressionHash]
-	if !ok {
-		return nil
-	}
-	// Lazy-expire: if the bucket is older than TTL, drop it and report
-	// "unknown". Keeps queries from returning stale peers after the
-	// dedup window has elapsed (which would feed the slot with
-	// counters the rest of the network would have dropped long ago).
-	if clock().Sub(entry.seenAt) >= RelayedIndexTTL {
-		delete(o.relayedIndex, suppressionHash)
-		return nil
-	}
-
-	out := make([]PeerID, 0, len(entry.peers))
-	for id := range entry.peers {
-		out = append(out, id)
-	}
-	return out
-}
-
 // OnValidatorMessage is called by the consensus router on every inbound
 // trusted proposal/validation so the reduce-relay state machine can
 // select peers to squelch.
@@ -2336,152 +1767,6 @@ func (o *Overlay) localLoadFeeProviderSnapshot() func() uint32 {
 // inclusion — entries reporting older than this are dropped before the
 // median is taken.
 const clusterFeeWindow = 90 * time.Second
-
-// PeersJSON implements types.PeerSource for the `peers` RPC method,
-// emitting the subset of rippled's per-peer RPC fields for which
-// go-xrpl has data.
-func (o *Overlay) PeersJSON() []map[string]any {
-	list := o.Peers()
-	out := make([]map[string]any, 0, len(list))
-	for _, p := range list {
-		entry := map[string]any{
-			"address":    p.Endpoint.String(),
-			"public_key": p.PublicKey,
-			"uptime":     int64(time.Since(p.ConnectedAt).Seconds()),
-			"load":       p.Load,
-		}
-		if p.Inbound {
-			entry["inbound"] = true
-		}
-		if p.ServerDomain != "" {
-			entry["server_domain"] = p.ServerDomain
-		}
-		// Emit only when the peer set a Network-ID.
-		if p.NetworkID != "" {
-			entry["network_id"] = p.NetworkID
-		}
-		if p.ClosedLedger != "" {
-			entry["ledger"] = p.ClosedLedger
-		}
-		if p.CompleteLedgers != "" {
-			entry["complete_ledgers"] = p.CompleteLedgers
-		}
-		if len(p.PublicKeyBytes) > 0 {
-			if member, ok := o.cluster.Member(p.PublicKeyBytes); ok {
-				entry["cluster"] = true
-				if member.Name != "" {
-					entry["name"] = member.Name
-				}
-			}
-		}
-		// Omit when converged.
-		switch p.Tracking {
-		case PeerTrackingDiverged:
-			entry["track"] = "diverged"
-		case PeerTrackingUnknown:
-			entry["track"] = "unknown"
-		}
-		if p.HasLatency {
-			entry["latency"] = uint32(p.Latency / time.Millisecond)
-		}
-		// Version sourced from User-Agent (inbound) or Server
-		// (outbound) header.
-		if p.Version != "" {
-			entry["version"] = p.Version
-		}
-		// Emit unconditionally — a negotiated value always exists once
-		// the handshake has completed.
-		entry["protocol"] = p.Protocol
-		// Emit only when the peer has reported a status.
-		if s, known := nodeStatusRPCName(p.Status); s != "" {
-			entry["status"] = s
-		} else if !known && p.Status != 0 {
-			// Log a warning when the status falls outside the known
-			// enum, then drop the field so out-of-range values aren't
-			// silent.
-			slog.Warn("Unknown peer status",
-				"t", "Overlay", "peer", p.ID, "status", int32(p.Status))
-		}
-		// Emit the metrics object; values are decimal strings to match
-		// rippled's formatting.
-		entry["metrics"] = map[string]any{
-			"total_bytes_recv": strconv.FormatUint(p.TotalBytesRecv, 10),
-			"total_bytes_sent": strconv.FormatUint(p.TotalBytesSent, 10),
-			"avg_bps_recv":     strconv.FormatUint(p.AvgBpsRecv, 10),
-			"avg_bps_sent":     strconv.FormatUint(p.AvgBpsSent, 10),
-			"send_drops":       strconv.FormatUint(p.SendDrops, 10),
-		}
-		out = append(out, entry)
-	}
-	return out
-}
-
-// nodeStatusRPCName returns the rippled spelling for each known
-// NodeStatus and a `known` flag distinguishing "no status reported"
-// (nsUNKNOWN, known=true) from "unrecognized enum value" (known=false).
-// The caller omits the `status` field for either case but logs only
-// the unknown-enum case.
-func nodeStatusRPCName(s message.NodeStatus) (string, bool) {
-	switch s {
-	case 0:
-		return "", true
-	case message.NodeStatusConnecting:
-		return "connecting", true
-	case message.NodeStatusConnected:
-		return "connected", true
-	case message.NodeStatusMonitoring:
-		return "monitoring", true
-	case message.NodeStatusValidating:
-		return "validating", true
-	case message.NodeStatusShutting:
-		return "shutting", true
-	default:
-		return "", false
-	}
-}
-
-// clusterFeeRef is the load-fee reference baseline. Replace with a
-// live reference once go-xrpl grows a load-fee tracker.
-const clusterFeeRef uint32 = 256
-
-// ClusterJSON returns the top-level cluster object for the `peers`
-// RPC response.
-func (o *Overlay) ClusterJSON() map[string]any {
-	out := map[string]any{}
-	if o == nil || o.cluster == nil {
-		return out
-	}
-
-	var selfKey []byte
-	if o.identity != nil {
-		selfKey = o.identity.PublicKey()
-	}
-
-	now := o.cfg.Clock()
-
-	o.cluster.ForEach(func(m cluster.Member) {
-		if len(selfKey) > 0 && bytes.Equal(selfKey, m.Identity) {
-			return
-		}
-		encoded, err := addresscodec.EncodeNodePublicKey(m.Identity)
-		if err != nil || encoded == "" {
-			return
-		}
-		entry := map[string]any{}
-		if m.Name != "" {
-			entry["tag"] = m.Name
-		}
-		if m.LoadFee != clusterFeeRef && m.LoadFee != 0 {
-			entry["fee"] = float64(m.LoadFee) / float64(clusterFeeRef)
-		}
-		if !m.ReportTime.IsZero() {
-			age := max(int64(now.Sub(m.ReportTime).Seconds()), 0)
-			entry["age"] = age
-		}
-		out[encoded] = entry
-	})
-	return out
-}
 
 // PeerCount returns the number of connected peers.
 func (o *Overlay) PeerCount() int {
